Add type alias and factor out partition rdd
